datastore - sanity refactor and greenlet bug fix
